make it work on linux and freebsd
