+++ pkg/client/cli.go
@@ -1,12 +1,14 @@
 package client
 
 import (
+	"bufio"
 	"bytes"
 	"context"
 	"fmt"
 	"io"
 	"math/rand"
 	"net/http"
+	"os"
 	"strings"
 	"time"
 
@@ -15,6 +17,7 @@ import (
 	"github.com/Notifiarr/notifiarr/pkg/logs"
 	"github.com/Notifiarr/notifiarr/pkg/mnd"
 	"github.com/Notifiarr/notifiarr/pkg/services"
+	"github.com/Notifiarr/notifiarr/pkg/website"
 	"golift.io/rotatorr"
 )
 
@@ -165,4 +168,61 @@ func printCurlReply(resp *http.Response, body []byte) {
 func Fortune() string {
 	fortunes := strings.Split(bindata.MustAssetString("other/fortunes.txt"), "\n%\n")
 	return fortunes[rand.Intn(len(fortunes))] //nolint:gosec
+}
+
+// handleAptHook takes a payload as stdin from dpkg and relays it to notifiarr.com.
+// only useful as an apt integration on Debian-based operating systems.
+// NEVER return an error, we don't want to hang up apt.
+func (c *Client) handleAptHook(ctx context.Context) error {
+	if !mnd.IsLinux {
+		return ErrUnsupport
+	} else if !c.Config.EnableApt {
+		return nil // apt integration is not enabled, bail.
+	}
+
+	var (
+		grab   bool
+		output struct {
+			Data    []string `json:"data"`
+			CLI     string   `json:"cli"`
+			Install int      `json:"install"`
+			Remove  int      `json:"remove"`
+		}
+	)
+
+	for scanner := bufio.NewScanner(os.Stdin); scanner.Scan(); {
+		switch line := scanner.Text(); {
+		case strings.HasPrefix(line, "CommandLine"):
+			output.CLI = line
+		case line == "":
+			grab = true // grab everything after the empty line.
+		case grab:
+			output.Data = append(output.Data, line)
+
+			if strings.HasSuffix(line, ".deb") {
+				output.Install++
+			} else if strings.HasSuffix(line, "**REMOVE**") {
+				output.Remove++
+			}
+
+			fallthrough
+		default: /* debug /**/
+			// fmt.Println("hook line", line)
+		} //nolint:wsl
+	}
+
+	resp, _, err := c.website.RawGetData(ctx, &website.Request{
+		Route:   website.PkgRoute,
+		Event:   "apt",
+		Payload: output,
+	})
+	//nolint:forbidigo
+	if err != nil {
+		fmt.Printf("ERROR Sending Notification to notifiarr.com: %v%s\n", err, resp)
+	} else {
+		fmt.Printf("Sent notification to notifiarr.com; install: %d, remove: %d%s\n",
+			output.Install, output.Remove, resp)
+	}
+
+	return nil
 }
