+++ pkg/ui/dlgs.go
@@ -1,4 +1,4 @@
-//go:build windows || darwin
+//go:build windows || darwin || linux
 
 package ui
